Moved verify_captcha to a core/general location as it can be used by other apps
